F Introduce a Wall arround the MirrorGrid in order to mirror the GridElements correctly

+++ MoveableImpl/src/com/myownb3/piranha/util/MathUtil.java
@@ -42,7 +42,7 @@ public class MathUtil {
     *        the offset
     * @return a random number
     */
-   public static double getRandom(int offset) {
+   public static double getRandom(double offset) {
       return Math.random() * offset;
    }
